Restructure the cobra commands
