+++ .gitignore
@@ -3,6 +3,7 @@ main
 cli-utils
 dyctl
 kapply
+bin
 
 #intellij
 .idea/
